requirements.txt
Django==1.11.2
tensorflow==1.1.0
Pillow==4.1.1
